completed all controllers

+++ src/controllers/leadCtrl.ts
@@ -1,12 +1,17 @@
 import { Request, Response } from "express";
 import Lead from "../models/leadModel";
 import Apartment from "../models/apartmentModel";
-import { IApartment, ILead, IResponse } from "../config/interfaces";
+import { IApartment, ILead, IRequest, IResponse } from "../config/interfaces";
 
 const leadCtrl = {
-  getLeads: async (req: Request, res: IResponse) => {
+  getLeads: async (req: IRequest, res: IResponse) => {
     try {
-      const leads = await Lead.find({});
+      const Allowed = req.isAllowed;
+      if (!Allowed) return res.error.notAllowed(res);
+
+      const leads = await Lead.find(
+        req.role === "superadmin" ? {} : { isShow: true }
+      );
 
       res.json({ status: "OK", length: leads.length, leads });
     } catch (err: any) {
@@ -43,8 +48,11 @@ const leadCtrl = {
       return res.error.handleError(res, err);
     }
   },
-  deleteLead: async (req: Request, res: IResponse) => {
+  deleteLead: async (req: IRequest, res: IResponse) => {
     try {
+      const Allowed = req.isAllowed;
+      if (!Allowed) return res.error.notAllowed(res);
+
       const lead = await Lead.findByIdAndDelete(req.params.id);
       if (!lead) return res.error.leadNotFound(res);
